Project_002-Create_port_configuration_based_on_Excel_File/requirements.txt
asttokens==3.0.1
bcrypt==5.0.0
cffi==2.0.0
cryptography==46.0.3
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.1
invoke==2.2.1
ipdb==0.13.13
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
netmiko==4.6.0
ntc_templates==8.1.0
numpy==2.4.0
openpyxl==3.1.5
pandas==2.3.3
paramiko==4.0.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.1
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rich==14.2.0
ruamel.yaml==0.18.17
ruamel.yaml.clib==0.2.15
scp==0.15.0
six==1.17.0
stack-data==0.6.3
textfsm==2.1.0
traitlets==5.14.3
tzdata==2025.3
wcwidth==0.2.14
